HealthMed/pyproject.toml
[tool.poetry]
name = "healthmed"
version = "0.1.0"
description = ""
authors = ["Alex"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
Django = "^4.2"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.0.2"}
psycopg2-binary = "^2.9.6"
django-stubs = "^1.16.0"
djangorestframework = "^3.14.0"
#argon2 = "^0.1.10"
bcrypt = "^4.0.1"
passlib = "^1.7.4"
djangorestframework-simplejwt = "^5.2.2"
pymysql = "^1.0.3"
email-validator = "^2.0.0.post2"
pillow = "^9.5.0"
django-admin-interface = "^0.25.0"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
